Added running of all test cases to build script, in target "unittest". Target "all" now runs "unittest" and "windows". Had to add a cast to double in SortOfflineTest because JUnit 1.6 expected it.

+++ test/sort/SortOfflineTest.java
@@ -71,7 +71,7 @@ public class SortOfflineTest {// NOPMD
 		assertEquals("Events sorted wasn't the same as expected.",
 				expectedEvents, session.getEventsSorted());
 		assertEquals("Area in histogram wasn't the same as expected.",
-				expectedEvents, neutronE.getArea());
+				(double) expectedEvents, neutronE.getArea());
 	}
 
 	/**
